feat: 테스트 편의성을 위해 PostValidatorFactory 클래스를 인터페이스로 변경, 테스트용 구현체 클래스 추가
